Update [#0003]
Update [#0003]

1. There's a serious problem in project SampleFQ, something wrong with
the UDP protocol. Practically speaking, when server start broadcasting,
client can't always receive the message from server. And more seriously,
the broadcasting didn't work in my network environment. Therefore this
bug will be fixed in a long run, not today or tomorrow.
2. Currently i'm trying to run two server-client system with UDP
protocal, while one is TCP-liked, and another is a
broadcast-listen-liked. A good news is that I don't thick coding is the
problem, more specifically, success is depending on the network
environment of my own. :(

Summary @ 2015-07-07 20:16:39 UTC+8