Update [#0013]
Update [#0013]

#愛衣ちゃん大勝利い、つづく#

1. Replace former icons.
2. Add two constant classics to Universal.Global, named Port and
VerMessage. Port contains all the network ports will be used in this
project. VerMessage is an alias of Verification Message. This class
contains all the network stand messages in this project.
3. Add Login function and Register function. the PublishClient can now
connect to the PublishServer while the PublishServer is online.
4. Bug Fixed. Formerly, PublishServer will not exit normally, because of
a running thread. By now, when administrator shut down the Server, it
will automaticly join all thread.
5. I proudly announce that all the problems of network using have been
solved.

Summary @ 2015-07-28 15:26:42 UTC+8